+++ src/Entity/Meta.php
@@ -160,4 +160,12 @@ class Meta
             + $this->getExtendedNamespaces()
             + $this->getImplementedNamespaces();
     }
+
+    /**
+     * @return array
+     */
+    public function toArray()
+    {
+        return get_object_vars($this);
+    }
 }